Separate commands from manim modules
