+++ git_sim/add.py
@@ -1,11 +1,9 @@
 import sys
 import git
 import manim as m
-import typer
 
 from typing import List
 
-from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
 from git_sim.settings import settings
 
@@ -82,14 +80,3 @@ class Add(GitSimBaseCommand):
                         firstColumnArrowMap[z] = m.Arrow(
                             stroke_width=3, color=self.fontColor
                         )
-
-
-def add(
-    files: List[str] = typer.Argument(
-        default=None,
-        help="The names of one or more files to add to Git's staging area",
-    )
-):
-    settings.hide_first_tag = True
-    scene = Add(files=files)
-    handle_animations(scene=scene)
