+++ git_sim/enums.py
@@ -1,0 +1,14 @@
+from enum import Enum
+
+
+class ResetMode(Enum):
+    DEFAULT = "mixed"
+    SOFT = "soft"
+    MIXED = "mixed"
+    HARD = "hard"
+
+
+class StashSubCommand(Enum):
+    POP = "pop"
+    APPLY = "apply"
+    PUSH = "push"
