+++ git_sim/restore.py
@@ -1,10 +1,8 @@
 import sys
 import manim as m
-import typer
 
 from typing import List
 
-from git_sim.animations import handle_animations
 from git_sim.git_sim_base_command import GitSimBaseCommand
 from git_sim.settings import settings
 
@@ -71,14 +69,3 @@ class Restore(GitSimBaseCommand):
                         firstColumnArrowMap[y.a_path] = m.Arrow(
                             stroke_width=3, color=self.fontColor
                         )
-
-
-def restore(
-    files: List[str] = typer.Argument(
-        default=None,
-        help="The names of one or more files to restore",
-    )
-):
-    settings.hide_first_tag = True
-    scene = Restore(files=files)
-    handle_animations(scene=scene)
